XML: Fix Windows compilation and free allocated memory

+++ scilab/modules/xml/src/cpp/XMLNodeList.cpp
@@ -47,7 +47,7 @@ namespace org_modules_xml
         for (xmlNode * cur = parent->children; cur; cur = cur->next)
         {
             xmlNodeDump(buffer, doc.getRealDocument(), cur, 0, 1);
-            xmlBufferAdd(buffer, (xmlChar *)"\n", strlen("\n"));
+            xmlBufferAdd(buffer, (xmlChar *)"\n", (int)strlen("\n"));
         }
         std::string str = std::string((const char *)buffer->content);
         xmlBufferFree(buffer);
@@ -172,7 +172,7 @@ namespace org_modules_xml
             setElementAtPosition(index, XMLElement(doc, list.getRealNode()));
             for (xmlNode * cur = list.getRealNode()->next; cur; cur = cur->next)
             {
-                setElementAtPosition((double)(pos++) + 0.5d, XMLElement(doc, cur));
+                setElementAtPosition((double)(pos++) + 0.5, XMLElement(doc, cur));
             }
         }
     }
